pyproject.toml
[tool.poetry]
name = "aws-vault"
version = "0.1.0"
description = ""
authors = ["Eduardo Diez"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hvac = "^2.2.0"
python-dotenv = "^1.0.1"
pymongo = "^4.7.3"
regex = "^2024.5.15"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
